# Improve speed of S3 file move operations