This PR is meant to speed up s3 file move operations in the ECS instance. 

Performance improvement is achieved with 2 main changes:
1. Switch from Threading to Multi-Processing
  -- Threading was previously required because lambda instances did not support multi-processing. Multi-processing processes are slightly more expensive to create, but more performant during run-time than threads. 
2. Use shared boto session object for each process. 
  -- Creating boto Session objects for every file move operation is very expensive for the number of file moves we are completing. Using the multi-processing Pool `initializer` keyword argument, we are able to create a Session at process initialization and use it for all move operations done during the life of the process. 

In my testing this update showed a minimum of a 10x improvement in s3 file moving speed. 

```
** Current Copy **
ingestion    | WARNING:root:1500 of 1500 files copied
ingestion    | WARNING:root:took 221.86 seconds, 0.147908 seconds per file

** Multi-Processing Copy **
ingestion    | WARNING:root:1500 of 1500 files copied
ingestion    | WARNING:root:took 17.16 seconds, 0.011437 seconds per file

ingestion    | WARNING:root:copy speed up: 12.93 x
```

Asana Task: https://app.asana.com/0/1203185253920158/1203312862152172