@mzappitello 2022-11-07
I'm a little concerned how this will pair with each converter also being on its own thread. we are already doing that though, and these wins speak for themselves.

LGTM